requirements.txt
numpy==1.19.2
opencv-python==4.4.0.44
pyvirtualcam==0.2.0
Pillow==7.2.0
google-cloud-vision==2.0.0
